+++ src/com/day3/oop/Truck.java
@@ -42,7 +42,7 @@ public class Truck extends Vehicle {
     }
 
     public Truck(long id, String model, int mileage, int manufactureYear, boolean insured,
-                 String engineType, int engineLitreSize, ArrayList<String> amenities) {
+                 String engineType, double engineLitreSize, ArrayList<String> amenities) {
         super(id, model, mileage, manufactureYear, insured);
         this.engineType = engineType;
         this.engineLitreSize = engineLitreSize;
